Adds resolve host mixin
