+++ lib/msf/core/post/dns.rb
@@ -1,0 +1,4 @@
+# -*- coding: binary -*-
+
+module Msf::Post::DNS
+end
